+++ examples/complete/aws/autoscale/main.tf
@@ -1,5 +1,6 @@
 provider "aws" {
   region = var.region
+  profile = var.profile
 }
 
 provider "hopsworksai" {