+++ examples/complete/aws/autoscale/variables.tf
@@ -2,3 +2,8 @@ variable "region" {
   type    = string
   default = "us-east-2"
 }
+
+variable "profile" {
+  type    = string
+  default = "default"
+}